added error functionality on bathroom that doesn't exist

+++ backend/src/db.js
@@ -150,12 +150,34 @@ export async function likeBathroom(userId, bathroomId) {
 }
 
 /**
+ * get bathroom
+ * @param {string} bathroomId bathroom id
+ * @returns {object | undefined} bathroom if exists
+ */
+async function getBathroom(bathroomId) {
+  const {rows} = await pool.query({
+    text: `
+      SELECT *
+      FROM bathrooms
+      WHERE id = $1
+    `,
+    values: [bathroomId],
+  });
+
+  return rows.length > 0 ? rows[0] : undefined;
+}
+
+/**
  * remove like from bathroom
  * @param {string} userId user id
  * @param {string} bathroomId bathroom id
- * @returns {void}
  */
 export async function unlikeBathroom(userId, bathroomId) {
+  if (!(await getBathroom(bathroomId))) {
+    // throw error if bathroom doesn't exist
+    throw (new Error('bathroom doesn\'t exist'));
+  }
+
   await pool.query({
     text: `
       DELETE FROM userLikes
